% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ahrf_county}
\alias{ahrf_county}
\title{AHRF County Data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3230 rows and 6963 columns.}
\usage{
ahrf_county
}
\description{
AHRF County Data
}
\keyword{datasets}

