% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ahrf_state_layout}
\alias{ahrf_state_layout}
\title{AHRF State Data Layout}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1122 rows and 6 columns.}
\usage{
ahrf_state_layout
}
\description{
AHRF State Data Layout
}
\keyword{datasets}

