% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ahrf_state}
\alias{ahrf_state}
\title{AHRF State Data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 52 rows and 1122 columns.}
\usage{
ahrf_state
}
\description{
AHRF State Data
}
\keyword{datasets}

